requirements.txt
Django
djangorestframework
psycopg2
celery
redis